refactor: deprecated 파일 별도 폴더에 관리
