+++ src/components/deprecated/StudyModal/StudyModal.tsx
